Resolving order of literals in sdata2
