+++ obj_files.mk
@@ -137,6 +137,7 @@ TEXT_O_FILES := 						            \
 	$(BUILD_DIR)/src/melee/ft/chara/ftZelda/ftzelda2.c.o 		\
 	$(BUILD_DIR)/src/melee/ft/chara/ftZelda/ftzelda3.c.o 		\
 	$(BUILD_DIR)/src/melee/ft/chara/ftZelda/ftzelda4.c.o 		\
+	$(BUILD_DIR)/src/melee/ft/chara/ftZelda/ftzelda5.c.o 		\
 	$(BUILD_DIR)/asm/melee/ft/chara/ftzelda.s.o 		\
 	$(BUILD_DIR)/src/melee/ft/chara/ftPurin/ftpurin.c.o 		\
 	$(BUILD_DIR)/asm/melee/ft/chara/ftpurin.s.o 		\
