+++ src/melee/ft/chara/ftZelda/ftzelda3.c
@@ -1,13 +1,13 @@
 #include "ftzelda.h"
 
-// 8013ADB4 - 8013AE30 (0x7C bytes)
-// https://decomp.me/scratch/LbMVE
-void ftZelda_8013ADB4(HSD_GObj* fighterObj) {
-    Fighter* fighter;
+// 8013A830 - 8013A8AC (0x7C bytes)
+// https://decomp.me/scratch/6v90P
+void ftZelda_8013A830(HSD_GObj* fighterObj) {
+    Fighter* fighter; // r31
 
     fighter = fighterObj->user_data;
-    if ((u8) fighter->x2219_flag.bits.b0 == 0) {
-        ef_Spawn(0x4FC, fighterObj, fighter->x5E8_fighterBones[0x68].x0_jobj);
+    if (fighter->x2219_flag.bits.b0 == 0) {
+        ef_Spawn(0x4F4, fighterObj, fighter->x5E8_fighterBones[1].x0_jobj);
         fighter->x2219_flag.bits.b0 = 1;
     }
     fighter->cb.x21D4_callback_EnterHitlag = efLib_PauseAll;
@@ -15,14 +15,14 @@ void ftZelda_8013ADB4(HSD_GObj* fighterObj) {
     fighter->cb.x21BC_callback_Accessory4 = 0;
 }
 
-// 8013AE30 - 8013AEAC (0x7C bytes)
-// https://decomp.me/scratch/iojLO
-void ftZelda_8013AE30(HSD_GObj* fighterObj) {
-    Fighter* fighter;
+// 8013A8AC - 8013A928 (0x7C bytes)
+// https://decomp.me/scratch/rfSO5
+void ftZelda_8013A8AC(HSD_GObj* fighterObj) {
+    Fighter* fighter; // r31
 
     fighter = fighterObj->user_data;
-    if ((u8) fighter->x2219_flag.bits.b0 == 0) {
-        ef_Spawn(0x4FD, fighterObj, fighter->x5E8_fighterBones[4].x0_jobj);
+    if (fighter->x2219_flag.bits.b0 == 0) {
+        ef_Spawn(0x4F5, fighterObj, fighter->x5E8_fighterBones[1].x0_jobj);
         fighter->x2219_flag.bits.b0 = 1;
     }
     fighter->cb.x21D4_callback_EnterHitlag = efLib_PauseAll;
@@ -30,300 +30,179 @@ void ftZelda_8013AE30(HSD_GObj* fighterObj) {
     fighter->cb.x21BC_callback_Accessory4 = 0;
 }
 
-// 8013AEAC - 8013AEE0 (0x34 bytes)
-// https://decomp.me/scratch/iINH1
-void func_80114758(HSD_GObj*);                      /* extern */
-void ftZelda_8013AEAC(HSD_GObj* fighterObj) {
-    Fighter* fighter;
-    fighter = fighterObj->user_data;
-    fighter->cb.x21BC_callback_Accessory4 = 0;
-    func_8007EFC8(fighterObj, func_80114758);
-}
-
-// 8013AEE0 - 8013AFA4 (0xC4 bytes)
-// https://decomp.me/scratch/Lw6fO
-void ftZelda_SpecialLw_StartAction(HSD_GObj* fighterObj) {
-    Point3d sp20;
+// 8013A928 - 8013A9A4 (0x7C bytes)
+// https://decomp.me/scratch/arJgt
+void ftZelda_SpecialN_StartAction(HSD_GObj* fighterObj) {
     f32 temp_f1;
     Fighter* fighter; // r31
-    ftZeldaAttributes* attributes; // r3
+    Fighter* fighter2; // r5
+    ftZeldaAttributes* attributes;
     f32 unused[5];
 
     temp_f1 = 0;
-    Fighter_ActionStateChange_800693AC(fighterObj, 0x163, 0, NULL, temp_f1, 1.0, temp_f1);
-    func_8006EBA4(fighterObj);
-    
     fighter = fighterObj->user_data;
-    attributes = fighter->x2D4_specialAttributes;
-    fighter->x2200_ftcmd_var0 = 0;
-    fighter->x80_self_vel.x = (f32) (fighter->x80_self_vel.x / attributes->x70);
-    fighter->x80_self_vel.y = (f32) (fighter->x80_self_vel.y / attributes->x74);
-    fighter->xEC_ground_vel = (f32) (fighter->xEC_ground_vel / attributes->x70);
-    
-    func_8000B1CC(fighter->x5E8_fighterBones[0].x0_jobj, NULL, &sp20);
-    func_800119DC(&sp20, 0x78, 0.4000000059604645, 0.003000000026077032, 1.0471975803375244);
-    
-    fighter->cb.x21BC_callback_Accessory4 = &ftZelda_8013ADB4;
+    Fighter_ActionStateChange_800693AC(fighterObj, 0x155, 0, NULL, temp_f1, 1.0, temp_f1);
+    func_8006EBA4(fighterObj);
+    fighter2 = fighterObj->user_data;
+    attributes = fighter2->x2D4_specialAttributes;
+    fighter2->x2200_ftcmd_var0 = 0;
+    fighter2->x2340_stateVar1 = attributes->x4;
+    fighter->cb.x21BC_callback_Accessory4 = &ftZelda_8013A830;
 }
 
-// 8013AFA4 - 8013B068 (0xC4 bytes)
-// https://decomp.me/scratch/8W7ZF
-void ftZelda_SpecialAirLw_StartAction(HSD_GObj* fighterObj) {
-    Point3d sp20;
-    f32 temp_f1;
+// 8013A9A4 - 8013AA38 (0x94 bytes)
+// https://decomp.me/scratch/DgO7D
+void ftZelda_SpecialAirN_StartAction(HSD_GObj* fighterObj) {
     Fighter* fighter; // r31
-    ftZeldaAttributes* attributes; // r3
+    Fighter* fighter2;
+    ftZeldaAttributes* attributes;
     f32 unused[5];
 
-    temp_f1 = 0;
-    Fighter_ActionStateChange_800693AC(fighterObj, 0x165, 0, NULL, temp_f1, 1.0, temp_f1);
+    fighter = fighterObj->user_data;
+    attributes = fighter->x2D4_specialAttributes;
+    fighter->x80_self_vel.y = 0;
+    fighter->x80_self_vel.x = fighter->x80_self_vel.x / attributes->x8;
+    Fighter_ActionStateChange_800693AC(fighterObj, 0x156, 0, NULL, 0, 1.0, 0);
     func_8006EBA4(fighterObj);
+    fighter2 = getFighter(fighterObj);
+    attributes = fighter2->x2D4_specialAttributes;
+    fighter2->x2200_ftcmd_var0 = 0;
+    fighter2->x2340_stateVar1 = attributes->x4;
+    fighter->cb.x21BC_callback_Accessory4 = &ftZelda_8013A8AC;
+}
+
+// 8013AA38 - 8013AACC (0x94 bytes)
+// https://decomp.me/scratch/XcH4E
+void ftZelda_8013AA38(HSD_GObj* fighterObj) {
+    Fighter* fighter;
+    s32 attributes;
+    s32 unused[5];
+
     fighter = fighterObj->user_data;
-    attributes = fighter->x2D4_specialAttributes;
-    fighter->x2200_ftcmd_var0 = 0;
-    fighter->x80_self_vel.x = (f32) (fighter->x80_self_vel.x / attributes->x70);
-    fighter->x80_self_vel.y = (f32) (fighter->x80_self_vel.y / attributes->x74);
-    fighter->xEC_ground_vel = (f32) (fighter->xEC_ground_vel / attributes->x70);
+    attributes = ((u32)fighter->x2D4_specialAttributes);
     
-    func_8000B1CC(fighter->x5E8_fighterBones[0].x0_jobj, NULL, &sp20);
-    func_800119DC(&sp20, 0x78, 0.4000000059604645, 0.003000000026077032, 1.0471975803375244);
-    fighter->cb.x21BC_callback_Accessory4 = &ftZelda_8013ADB4;
-}
-
-// 8013B068 - 8013B0A8 (0x40 bytes)
-// https://decomp.me/scratch/DrRr5
-void ftZelda_8013B068(HSD_GObj* fighterObj) {
-    Fighter* fighter;
-
-    fighter = fighterObj->user_data;
+    if ((u32) fighter->x2200_ftcmd_var0 == 1) {
+        fighter->x2200_ftcmd_var0 = 2;
+        ftColl_CreateReflectHit(fighterObj, attributes + 0x84, &ftZelda_8013ADB0);
+    }
+    if ((u32) fighter->x2200_ftcmd_var0 == 0) {
+        fighter->x2218_flag.bits.b3 = 0;
+    }
     if (ftAnim_IsFramesRemaining(fighterObj) == 0) {
-        fighter->cb.x21BC_callback_Accessory4 = &ftZelda_8013AEAC;
+        func_8008A2BC(fighterObj);
     }
 }
 
-// 8013B0A8 - 8013B0E8 (0x40 bytes)
-// https://decomp.me/scratch/DrRr5
-void ftZelda_8013B0A8(HSD_GObj* fighterObj) {
+// 8013AACC - 8013AB60 (0x94 bytes)
+// https://decomp.me/scratch/ttWvN
+void ftZelda_8013AACC(HSD_GObj* fighterObj) {
     Fighter* fighter;
+    s32 attributes;
+    s32 unused[5];
 
     fighter = fighterObj->user_data;
-    if (ftAnim_IsFramesRemaining(fighterObj) == 0) {
-        fighter->cb.x21BC_callback_Accessory4 = &ftZelda_8013AEAC;
-    }
-}
-
-void ftZelda_8013B0E8(HSD_GObj* arg0) {}
-void ftZelda_8013B0EC(HSD_GObj* arg0) {}
-
-// 8013B0F0 - 8013B110 (0x20 bytes)
-void ftZelda_8013B0F0(HSD_GObj* arg0) {
-    func_80084F3C(arg0);
-}
-
-// 8013B110 - 8013B154 (0x44 bytes)
-// https://decomp.me/scratch/Nm958
-void ftZelda_8013B110(HSD_GObj* fighterObj) {
-    Fighter* fighter;
-    ftZeldaAttributes* attributes;
-    f32 unused[3];
-
-    fighter = fighterObj->user_data;
-    attributes = fighter->x2D4_specialAttributes;
-    func_8007D494(fighter, attributes->x78, attributes->x7C);
-    func_8007CEF4(fighter);
-}
-
-// 8013B154 - 8013B190 (0x3C bytes)
-// https://decomp.me/scratch/8ceOO
-void ftZelda_8013B154(HSD_GObj* fighterObj) {
-    if (func_80082708(fighterObj) == 0) {
-        ftZelda_8013B1CC(fighterObj);
-    }
-}
-
-// 8013B190 - 8013B1CC (0x3C bytes)
-// https://decomp.me/scratch/8ceOO
-void ftZelda_8013B190(HSD_GObj* fighterObj) {
-    if (func_80081D0C(fighterObj) != 0) {
-        ftZelda_8013B238(fighterObj);
-    }
-}
-
-// 8013B1CC - 8013B238 (0x6C bytes)
-// https://decomp.me/scratch/w04qW
-void ftZelda_8013B1CC(HSD_GObj* fighterObj) {
-    Fighter* fighter;
-
-    fighter = fighterObj->user_data;
-    func_8007D5D4(fighter);
-    Fighter_ActionStateChange_800693AC(fighterObj, 0x165, 0x0C4C508E, NULL, fighter->x894_currentAnimFrame, 1.0, 0);
-    fighter->cb.x21BC_callback_Accessory4 = &ftZelda_8013ADB4;
-}
-
-// 8013B238 - 8013B2A4 (0x6C bytes)
-void ftZelda_8013B238(HSD_GObj* fighterObj) {
-    Fighter* fighter;
-
-    fighter = fighterObj->user_data;
-    func_8007D7FC(fighter);
-    Fighter_ActionStateChange_800693AC(fighterObj, 0x163, 0x0C4C508E, NULL, fighter->x894_currentAnimFrame, 1.0, 0);
-    fighter->cb.x21BC_callback_Accessory4 = &ftZelda_8013ADB4;
-}
-
-// 8013B2A4 - 8013B2E0 (0x3C bytes)
-void ftZelda_8013B2A4(HSD_GObj* arg0) {
-    if (ftAnim_IsFramesRemaining(arg0) == 0) {
-        func_8008A2BC(arg0);
-    }
-}
-
-// 8013B2E0 - 8013B31C (0x3C bytes)
-void ftZelda_8013B2E0(HSD_GObj* arg0) {
-    if (ftAnim_IsFramesRemaining(arg0) == 0) {
-        func_800CC730(arg0);
-    }
-}
-
-void ftZelda_8013B31C(HSD_GObj* arg0) {}
-void ftZelda_8013B320(HSD_GObj* arg0) {}
-
-// 8013B324 - 8013B344 (0x20 bytes)
-void ftZelda_8013B324(HSD_GObj* arg0) {
-    func_80084F3C(arg0);
-}
-
-// 8013B344 - 8013B388 (0x44 bytes)
-// https://decomp.me/scratch/r3T6y
-void ftZelda_8013B344(HSD_GObj* fighterObj) {
-    Fighter* fighter;
-    ftZeldaAttributes* attributes;
-    f32 unused[3];
-
-    fighter = fighterObj->user_data;
-    attributes = fighter->x2D4_specialAttributes;
-    func_8007D494(fighter, attributes->x78, attributes->x7C);
-    func_8007CEF4(fighter);
-}
-
-// 8013B388 - 8013B3C4 (0x3C bytes)
-// https://decomp.me/scratch/t8hrY
-void ftZelda_8013B388(HSD_GObj* fighterObj) {
-    if (func_80082708(fighterObj) == 0) {
-        ftZelda_8013B400(fighterObj);
-    }
-}
-
-// 8013B3C4 - 8013B400 (0x3C bytes)
-// https://decomp.me/scratch/t8hrY
-void ftZelda_8013B3C4(HSD_GObj* fighterObj) {
-    if (func_80081D0C(fighterObj) != 0) {
-        ftZelda_8013B46C(fighterObj);
-    }
-}
-
-// 8013B400 - 8013B46C (0x6C bytes)
-// https://decomp.me/scratch/EZlpO
-void ftZelda_8013B400(HSD_GObj* fighterObj) {
-    Fighter* fighter;
-
-    fighter = fighterObj->user_data;
-    func_8007D5D4(fighter);
-    Fighter_ActionStateChange_800693AC(fighterObj, 0x166, 0x0C4C508E, NULL, fighter->x894_currentAnimFrame, 1.0, 0);
-    fighter->cb.x21BC_callback_Accessory4 = &ftZelda_8013AE30;
-}
-
-// 8013B46C - 8013B4D8 (0x6C bytes)
-// https://decomp.me/scratch/EZlpO
-void ftZelda_8013B46C(HSD_GObj* fighterObj) {
-    Fighter* fighter;
-
-    fighter = fighterObj->user_data;
-    func_8007D7FC(fighter);
-    Fighter_ActionStateChange_800693AC(fighterObj, 0x164, 0x0C4C508E, NULL, fighter->x894_currentAnimFrame, 1.0, 0);
-    fighter->cb.x21BC_callback_Accessory4 = &ftZelda_8013AE30;
-}
-
-// 8013B4D8 - 8013B540 (0x68 bytes)
-// https://decomp.me/scratch/wpEbJ
-void ftZelda_8013B4D8(HSD_GObj* fighterObj) {
-    Fighter* fighter;
-    ftZeldaAttributes* attributes;
-    s32 var_r4;
-    s32 unused[1];
-
-    fighter = getFighter(fighterObj);
-    attributes = fighter->x2D4_specialAttributes;
+    attributes = ((u32)fighter->x2D4_specialAttributes);
     
-    if (fighter->xE0_ground_or_air == 0) {
-        var_r4 = 0x164;
+    if ((u32) fighter->x2200_ftcmd_var0 == 1U) {
+        fighter->x2200_ftcmd_var0 = 2U;
+        ftColl_CreateReflectHit(fighterObj, attributes + 0x84, &ftZelda_8013ADB0);
+    }
+    if ((u32) fighter->x2200_ftcmd_var0 == 0U) {
+        fighter->x2218_flag.bits.b3 = 0;
+    }
+    if (ftAnim_IsFramesRemaining(fighterObj) == 0) {
+        func_800CC730(fighterObj);
+    }
+}
+
+void ftZelda_8013AB60(HSD_GObj* fighterObj) {}
+void ftZelda_8013AB64(HSD_GObj* fighterObj) {}
+
+// 8013AB68 - 8013AB9C (0x34 bytes)
+void ftZelda_8013AB68(HSD_GObj* arg0) {
+    func_80084F3C(arg0);
+    func_8007AEF8(arg0);
+}
+
+// 8013AB9C - 8013AC10 (0x74 bytes)
+// https://decomp.me/scratch/RUgWL
+void ftZelda_8013AB9C(HSD_GObj* fighterObj) {
+    Fighter* fighter;
+    ftZeldaAttributes* charAttr;
+    s32 stateVar1;
+    f32 arg2, arg3;
+    attr* fighterAttr;
+    s32 unused[2];
+
+    fighter = fighterObj->user_data;
+    charAttr = fighter->x2D4_specialAttributes;
+    stateVar1 = fighter->x2340_stateVar1;
+    fighterAttr = &fighter->x110_attr;
+    
+    if (stateVar1 != 0) {
+        fighter->x2340_stateVar1 = stateVar1 - 1;
     } else {
-        var_r4 = 0x166;
+        arg2 = charAttr->xC;
+        arg3 = fighterAttr->x170_TerminalVelocity;
+        func_8007D494(fighter, arg2, arg3) ;
     }
-    Fighter_ActionStateChange_800693AC(fighterObj, var_r4, 0, NULL, attributes->x80, 1.0, 0);
-    fighter->cb.x21BC_callback_Accessory4 = &ftZelda_8013AE30;
+    func_8007CF58(fighter);
+    func_8007AEF8(fighterObj);
 }
 
-// 8013B540 - 8013B574 (0x34 bytes)
-// https://decomp.me/scratch/L6UUI
-s32 ftZelda_8013B540(HSD_GObj* fighterObj) {
-    s32 actionStateIndex;
+// 8013AC10 - 8013AC4C (0x3C bytes)
+// https://decomp.me/scratch/CT7dz
+void ftZelda_8013AC10(HSD_GObj* fighterObj) {
+    if (func_80082708(fighterObj) == 0) {
+        ftZelda_8013AC88(fighterObj);
+    }
+}
+
+// 8013AC4C - 8013AC88 (0x3C bytes)
+// https://decomp.me/scratch/CT7dz
+void ftZelda_8013AC4C(HSD_GObj* fighterObj) {
+    if (func_80081D0C(fighterObj) != 0) {
+        ftZelda_8013AD1C(fighterObj);
+    }
+}
+
+// 8013AC88 - 8013AD1C (0x94 bytes)
+// https://decomp.me/scratch/B9loR
+void ftZelda_8013AC88(HSD_GObj* fighterObj) {
+    s32 attributes;
     Fighter* fighter;
-
-    fighter = fighterObj->user_data;
-    actionStateIndex = fighter->x10_action_state_index;
-    if (((actionStateIndex == 0x158) || (actionStateIndex == 0x15B)) && (fighter->sa.zelda.x222C != 0U)) {
-        return 1;
-    }
-    return 0;
-}
-
-// 8013B574 - 8013B5C4 (0x50 bytes)
-// https://decomp.me/scratch/tgIRc (fork of below)
-// https://decomp.me/scratch/pTAiQ
-s32 ftZelda_8013B574(HSD_GObj* fighterObj) {
-    s32 temp_r0;
-    Fighter* fighter; // r3
+    Fighter* fighter2;
+    s32 unused[5];
 
     fighter = fighterObj->user_data;
     
-    if ((u32) fighter->sa.zelda.x222C != 0U) {
-        switch(fighter->x10_action_state_index) {
-            case 0x15C:
-            case 0x159:
-                if (fighter->x2204_ftcmd_var1 == 1) {
-                    fighter->x2204_ftcmd_var1 = 0;
-                    return 1;
-                }
-            break;
-        }
+    func_8007D5D4(fighter);
+    Fighter_ActionStateChange_800693AC(fighterObj, 0x156, 0x0C4C508E, NULL, fighter->x894_currentAnimFrame, 1.0, 0);
+    fighter2 = fighterObj->user_data;
+    attributes = ((u32)fighter2->x2D4_specialAttributes);
+    if ((u32) fighter2->x2200_ftcmd_var0 == 2U) {
+        ftColl_CreateReflectHit(fighterObj, attributes + 0x84, &ftZelda_8013ADB0);
     }
-    
-    return 0;
+    fighter->cb.x21BC_callback_Accessory4 = &ftZelda_8013A8AC;
 }
 
-// 8013B5C4 - 8013B5EC ( bytes)
-// https://decomp.me/scratch/VdBYJ
-void ftZelda_8013B5C4(HSD_GObj* fighterObj) {
+// 8013AD1C - 8013ADB0 (0x94 bytes)
+// https://decomp.me/scratch/pQ6Xn
+void ftZelda_8013AD1C(HSD_GObj* fighterObj) {
+    s32 attributes;
     Fighter* fighter;
+    Fighter* fighter2;
+    s32 unused[5];
 
     fighter = fighterObj->user_data;
-    if ((u32) fighter->sa.zelda.x222C != 0U) {
-        fighter->sa.zelda.x222C = 0U;
+    func_8007D7FC(fighter);
+    Fighter_ActionStateChange_800693AC(fighterObj, 0x155, 0x0C4C508E, NULL, fighter->x894_currentAnimFrame, 1.0, 0);
+    fighter2 = fighterObj->user_data;
+    attributes = ((u32)fighter2->x2D4_specialAttributes);
+    if ((u32) fighter2->x2200_ftcmd_var0 == 2U) {
+        ftColl_CreateReflectHit(fighterObj, attributes + 0x84, &ftZelda_8013ADB0);
     }
-    fighter->cb.x21E4_callback_OnDeath2 = 0;
-    fighter->cb.x21DC_callback_OnTakeDamage = 0;
+    fighter->cb.x21BC_callback_Accessory4 = &ftZelda_8013A830;
 }
 
-// 8013B5EC - 8013B638 (0x4C bytes)
-// https://decomp.me/scratch/8QoCa
-void ftZelda_8013B5EC(HSD_GObj* fighterObj) {
-    Fighter* fighter;
-
-    fighter = fighterObj->user_data;
-    if ((u32) fighter->sa.zelda.x222C != 0) {
-        func_802C3D44(fighter->sa.zelda.x222C);
-        fighter->sa.zelda.x222C = 0;
-    }
-    fighter->cb.x21E4_callback_OnDeath2 = 0;
-    fighter->cb.x21DC_callback_OnTakeDamage = 0;
-}
+void ftZelda_8013ADB0(HSD_GObj* fighterObj) {}
